Refactor: Migrate AppCubit from enum to freezed states
This commit refactors the `AppCubit` by replacing the `AppStatus` enum with freezed states (`AppState`). This change improves type safety and allows for more expressive state management.

The `MainView` has been updated to use the new `AppState` and its `whenOrNull` method for handling the unauthenticated state. Additionally, a logout button has been added to the `AppBar` in `MainView` to trigger the `logout` method in `AppCubit`.