Add proguard rules for Flutter Android apps
This commit introduces a `proguard-rules.pro` file to the `android/app` directory. This file contains rules to ensure that essential Flutter classes are not removed or obfuscated during the Proguard optimization process.

Specifically, it keeps classes related to:
- `io.flutter.app`
- `io.flutter.plugin`
- `io.flutter.util`
- `io.flutter.view`
- General `io.flutter`
- `io.flutter.plugins`
- `io.flutter.embedding.engine.plugins`
- `io.flutter.embedding.engine.renderer`
- `io.flutter.embedding.engine.systemchannels`

It also adds a directive to not warn about `io.flutter.embedding` and to generally ignore warnings.